requirements.txt
aiogram==2.25.1
aiomysql==0.1.1
pandas==2.0.3
python-dateutil==2.8.2
asyncio==3.4.3
openpyxl==3.1.2
python-dotenv~=1.0.1